rhtlputil.toYices: create directory if target file requires it.
hence, same behavior as in rhtlp.toJTLVInput, which creates
subdirectories for smv and spec files as needed. If verbosity is set
to positive value, a warning message is printed.

This error was discovered because examples/autonomous_car_road.py will
fail if the directory tmpspec does not already exist (e.g., in a fresh
source distribution).