steps toward read/write of RHTLPProb instances in XML; does *not* yet handle short-horizon problems.
...example usage in autonomous_car_road_xml.py.  To facilitate this
work, I've tried to make handling of cont_props (i.e. ``continuous
propositions'') more consistent in the rhtlp module.  To that end
RHTLPProb now has a method, getContProps, to obtain an internally
saved copy of these.  Further, we require cont_props to be a
dictionary, even if the values are None or nil polytopes.