Add checking for depedencies and optional packages to setup.py.
This is custom code, i.e. does not require Distutils, PIP, etc.  The
checking mechanism is written to support easily adding/removing
dependencies; see comments in the source file to that end.  To perform
a check, execute

python setup.py check

If "install" is detected in the argument list when invoking setup.py,
the checking procedure executes before installation of TuLiP.