remove unused function argument; work-around alignment bug for Matlab imports.
To see the bug, switch to the previous revision and change the flag
"use_mpt" to True in examples/robot_simple_disturbance.py

When reading data from Matlab MAT files, apparently scipy.io.loadmat
does not ensure correct alignment for the platform on which the code
is running.  Copying the matrix does *not* fix this, even though newly
created matrices have the ALIGNED flag set. Misalignment is only
problematic in certain cases, in particular for LAPACK.

The easiest work-around I have found is to export the ndarray to a
list of double (i.e. 64-bit floating point) elements and then create a
new ndarray from this.  There is no data loss because we use the
numpy.float64 data-type for polytopes.