synchronous product commutativity, special funcs, export formatting
export: added (custom) formatting support for label formatting

synchronous product between TS and BA: argument order now matters.
=================================================================

Motivated by a question during the meeting on the commutativity (in code)
of the synchronous product and the function name, and also by the need
to write the code for computing the BA representation of the
product between TS and BA, the following discusses the reasons
for the relevant design decisions.

The class (TS vs BA) of the sync product was defined with fixed
argument order, taking (TS, BA) and returning (prodTS, BA).
It has been changed to private, not meant for users any more.

Conceptually it should be a function, because it operates between
two different classes, so it does not naturally belong to any of
them. Nonetheless, a way to accomodate for whether the result
will be a (TS, persistent_set) or a BA is needed.

One way would be to create a second function, with a different name,
which takes (TS, BA) or possibly (BA, TS) and returns their
synchronous product, but now represented as a BA instead of
a tuple (TS, persistent_set).

There are a number of issues with the above:

1) it is not a structurally elegant solution.

2) it is quite cumbersome to use, because the user has to
guess the name (one solution is reading the doc, but I don't like
to have to search through the doc before using new packages,
typing . and using arrow keys should be self-explanatory).

3) the function naming is quite ambiguous and cannot be
effortlessly remembered (e.g. ts_ba_sync_prod, vs
ts_ba_synchronous_product vs ba_ts_sync_product vs ...).

4) the textbook def of sync product is commutative,
so the argument order can be fixed to be arbitrary.
But in the code implementation there is some extra freedom.
Exploiting the argument order can be used to some advantage.

5) Operator overloading or new infix operators could be defined,
but using them requires reading the doc for the order semantics
and two different operators would be needed (for BA and TS representations),
if the products are considered in the code as commutative.

--

Due to the above, for the user, the synchronous product is now implemented
by methods within the TS and BA classes. These methods do call the
module's function noted above, but the user is no more required to
know about its existence and the developers are not constrained to keep
it where it is or with the same name.

The new methods have been defined with the Kronecker product in mind.
Incidentally, the Kronecker product is (usually) denoted as the
synchronous product: http://en.wikipedia.org/wiki/Kronecker_product

Considering the Kronecker product A x B, the result "encapsulates" B
"within" A, in the sense that B is "inserted" in each element of A
and then some appropriate binary operation applies (scalar-matrix
multiplication in the case of matrices). The resulting object
can be viewed as A with "extra" internal structure.
Of course, this viewpoint is subjective, but it could serve
as a way of remembering how the software is structured.

Motivated by this analogy, (for the user) the synchronous product is now
implemented by two methods, one within TS and another within BA.
Both return the synchronous product between the TS and the BA,
but represented in a different way. The two alternative
representations are as a TS, or as a BA.

The selected one is determined by the class of the first argument of
the respective operation. In each case, the first argument is
the object to which the method belongs. This definition is
implicit, i.e., while programming the only argument given
to the method is the second argument of the binary operation.

These semantics are unambiguous and quite obvious, so the user
doesn't have to remember anything about order of arguments.
Secondly, the method's name is again straightforward: sync_prod().
The result is:
	prod_TS = TS.sync_prod(BA)
	prod_BA = BA.sync_pord(TS)

Besides these gains, the method can serve a dual purpose,
taking synchronous products between TS_1 x TS_2 and BA_1 x BA_2.
In the first case the argument given will be TS_2, so the
command becomes:
	prod_TS = TS1.sync_prod(TS_2)
and in the second case:
	prod_BA = BA1.sync_prod(BA_2)
Notice that the class of the result is naturally derived from
the previous definition, i.e., the first argument defines the class
(TS_1, BA_1, respectively). In this case the result would be
the same if the argument order was swapped, by issuing:
	prod_TS = TS2.sync_prod(TS_1)
	prod_BA = BA2.sync_prod(TS_1)
This is in accord with the commutativity which is expected by
multiplying TSs and BAs among themselves (symmetry).
Also, the name of the method remains the same, so we don't
need to define a new method.

The synchronous product is now also implemented by the arithmetic
operators __mul__ and the asynchronous product will be implemented
by __or__. These can exist together with the sync_prod() method.
So:
	TS.sync_prod(BA) = TS *BA
	BA.sync_prod(TS) = BA *TS
type(TS*BA) = FTS
type(BA*TS) = BA
and:
	TS1.sync_prod(TS2) = TS2.sync_prod(TS1) = TS1 *TS2 = TS2 *TS1
	BA1.sync_prod(BA2) = BA2.sync_prod(BA1) = BA1 *BA2 = BA2 *BA1
type(TS1*TS2) = FTS
type(BA1*BA2) = BA