1) implemented mutable states (unhashable) Extensive changes internally. Option transys.FTS(mutable=True) uses mutable states, so that unhashable objects can be stored.
This enables using any system in transys for algorithms which
do state refinement, during which it is convenient to store
the temporary states themselves in the graph.
For example, when parsing a language, or converting LTL->BA,
an intermediate state might be ['a', '||', 'b'], which is not hashable.
After the refinement, this might be separated into 2 states: 'a', 'b'.

2) General set test between iterables,
to allow for lists or mix of lists with sets.

3) Implemented removed_state_callback from states object to parent system.

Whether final states should be detached as methods of automata, or
within the states object is debatable.

Firstly, accessing fts.states.initial, fts.states.add_initial,
ba.states.initial, ba.states.add_intiial, but ba.add_final_state
is awkward. Why should final states be separated in the system?
Does this not duplicate methods for accessing subsets of states ?

This is the case for BA, but DRA don't have the same type of final states
(they are better represented as pairs), while FTS don't have final
states at all. So it looks like final states are a special semantic
property of automata, whereas initial states are used by every system
to define reachability in a unified way.

Also, it appears that when a state is removed from within states,
then States should access the parent's final states and remove it also
from there for consistency (which is how more thinking about this was
forced in the first place). This violates the separation of classes.
It can be avoided via the deletion callback to the parent.

However, the decision is made in this way for the following reasons.

States have a single subset: initial states.
The reason is that initial states are common between all types of finite
labeled state digraphs. This is not true for final states.

Therefore, if final states are implemented within States, then
they would be available to FTS, which is wrong.

So a subclass States_with_final_States would be needed.
But then FTS, oFTS, FSM would be inherited from a LabeledStateDiGraph
with States and the FSA from a different object, which
undermines the inheritance hierarchy.

In order to avoid this, there is the alternative of creating States
such that it provides any desired number of named subsets.
FTS can initialize this to initial states, whereas an FSA would
crete two distinct subsets: initial and final states.

At first this appears as an attractive option, but it is not.
Consider final states of a DRA. They are different, better stored as 2-tuples.
But then the subset provision breaks and more specialized methods are
needed from States.

This would lead States to become more and more compelx, with customary
methods implemented to serve each possible child of a LabeledStateDiGraph.
In turn, this doesn't make sense. Instead of ending like this,
why not implement a single huge object in the first place ?

For the above reasons functionality for final states is specialized to
subclasses and should not be placed within States.
Each subclass should implement its own methods for managing of final states
and their storage.

However, when states are removed from States, the parent (system)
should be notified to take care of any information it maintains
itself related to states, of which final states are one particular example.
This functionality is provided by the new callback.