-closed FTS export to Promela
-erased some remnant edge manipulation methods from FSA.
Moved them to LabeledTransitions, where they caused an
upheaval and significant rewrite of transition filtering
methods, end-result: single find transition command
for both searching for labels of transitions between selected
state pairs, as well as searching the preimage under edge labeling
of a set of selected edge sublabel valuations.

-edge keys abolished: not visible to the user any more !
This is to enforce uniformity, relieve both users and developers
from unnecessary and confusing double calls of the type:
"give me this, which I will give back to another method of yours...
so that I eventually get what I want..."

-never assign to .edge, and use only methods or  getitem magic
of networkx. This will allow us to implement a fork of networkx
with the new efficient storage and the same methods for interface.

-pulled str2singleton out of tuple2fts

-updated basic examples