removed AtomicPropositions, reimplemented selected methods in Labeled States:
AtomicPropositions class deleted.
FTS.atomic_propositions are now the MathSet of the PowerSet which is used to check state sublabel consistency.

This unifies label management for transition systems, automata and machines.
The labelings are defined in each type of system by a dict, which informs the checking methods about the sublabel types.

Reimplemented and merged methods in LabeledStates.
A single find method serves all search purposes in LabeledStates (i.e., you want to find the labels of selected states, or you want to find all states with a desired label, or both constraints superimposed).

LabeledStates.labels is intended as a plural for label (as a verb, denoting action), preferred for now vs longer names in the sake of simplicity. Otherwise the number of methods will multiply w/o reason, so the user will be forced to read docs, instead of deciding between just 2 available method names.

Pulled out sublabel management methods to their own class (pure method contaner, no data) LabelConsistency. This provides a uniform mechanism for checking sublabels of both LabeledStates and LabeledTransitions.