solve_closed_loop: return when target polytope becomes empty
The previous version returned as soon as s0 became empty, if not using all the horizon.

This change adds the case that all the horizon is used, and both s0 and p2 are empty. Since the union s0 \cup p2 is recursively taken, this is possible if p2 is initialized empty. This is possible only by passing an empty P2 as argument.