new synth.synthesize arg: ignore_ts_init
To ignore which initial states an OpenFTS or FTS has.
Useful to mask the absence of initial states whenever this is intentional.
E.g., adding initial states to the TS is omitted if it is desired to specify them directly at the spec level (i.e., by "manually" writing to sys_init).