support transition systems with arbitrary states for gr1c
If a TS has states of the form: letter + int, then each is converted to an int equal to the int they contain. This can prove convenient for users to keep track of the states after conversion to gr1c format.

If a TS has arbitrary states, then they are mapped to ints.
Note that a double conversion is implemented:
    1. synth.states2ints represents the TS in spec form using the state names as they are
    2. then GRSpec.to_gr1c is responsible for replacing them with ints and gr1cint.load_aut_xml for defining the proper types for the Mealy machine produced.

This might be somewhat inefficient (as opposed to using an int-valued state directly).
Reasons are:
    - This way the final machine has the states defined by the user, instead of the ints used by gr1c during solution.
    - If any state is referenced in the manually written spec part, e.g., loc = a, then the symbol 'a' will be properly replaced.