update examples for actions represented by int var
If more than 2 actions are defined, then by default mutual exclusion is assumed.
So an int variable models them in GR(1).
If they appear in the spec, they should be in one of the following forms:

    - act = action_value
    - eact = action_value

depending on whether a sys or env action is involved.

This requires that the user anticipate internal behavior of synth.
To be avoided by replacing lone occurrences of 'action_value' in the logic part of the spec by 'act = action_value' or 'eact = action_value', as appropriate.