rewrite AbstractSysDyn
So that it stores only a single mode, to avoid attributes which may or may not be dicts.
Add more info: storing PPPs instead of (unstructured) lists of Regions.
This should help keep track of abstraction info easier, hierarchically.
Also store both the partition which preserves both PWA dynamics and propositions, as well as the original PPP, which involves only the continuous propositions.