return (j, subsystem) from AbstractSwitched.ppp2sys
which needs storing the PwaSysDyn (always PWA for switched systems, will be LtiSysDyn if that is passed to abstract.discretize) in the AbstractPwa, so that it can retrieve the subsystem.
This bundles all information together and ensures correct access to data, avoiding error-prone carrying around of separate AbstractPwa and PwaSysDyn objects.
It should also make life easier when exporting abstractions.