bugfix: correct edge orientation produced by discretize
The original discretization and merging code follows an edge orientation convention opposite to standard convention.
The result was every edge being reverted when creating a transition system from a scipy.sparse.lil_matrix using ts.transitions.add_adj, because it calls nx.from_scipy_sparse_matrix, which follows the standard convention.

Indices were also reversed in plot_partition, yielding correct plots. So the edges were reverted only in the TS and were reverted back when plotting, falsly appearing as correct, unless one compared str(TS) with the plot.

Moreover, the inconsistency was hidden in synthesis of simple examples, because liveness requires edges in both directions, reducing the likelihood of exposing this.

It was revealed by the double_tank example, because there 'normal' mode drains the tanks, whereas 'refuel' mode fills them and 'refuel' is triggered only when the tanks are low, so the problem is asymmetric. The merging code reversed the transition directions, leading to the tanks filling in 'normal' mode (!), which then never triggered 'refuel' mode and led to an acyclic result.