improve state conversion to integers
If states are of the form 'x0',... where x is any character, then try mapping them to the integer part of the string.
Check that if N states are given, then the integers contained in them are from 0:N-1.
If not, so some integers are missing, fall back to using an arbitrary finite domain for modeling the states.