bugfix: correct use_all_horizon
which was using the running union of presets to feed it as the new preset for the next step in the iteration over time (up to the 0, starting from the horizon).
This yields erroneous results if drift, control constraints (Uset) or dynamics (A) imply that control cannot keep the system stationary.

The correct version keeps a running union of presets, but feeds back to the iteration the preset resulting from the previous iteration.