label types can be extended only explicitly from now on,
not during addition of edges.

This avoids the problematic design of adding new label values to the label typing object the user has passed when instantiating the LabeledStateDiGraph, by requiring that the user explicitly first (collects if necessary, e.g., using list comprehension), then adds those values to the label type, before adding the labeled edges.

From now on checks are performed only in TypedDict, by calling __contains__ of the type. Machine guard mechanisms will have to be updated to this simple approach.