use empty sets to represent the empty AP subset
value None should be reserved for special use and also is not compatible with the set container of atomic propositions (for labels that are non-empty sets).

tuple2fts converts None labels to empty sets, because they are more convenient for manual typing (and clear to read). Note that it does not deviate from the previous comment, because None is used as a shorthand while specifying inputs to tuple2fts, which is the usual Python convention to use the default value (here it is set()).