mv version to tulip/version.py
to avoid loading the tulip package when importing the version in setup.py, because it occurs before dependencies are installed.
This doesn't currently cause problems because __init__ defines __all__, instead of importing specific modules and sub packages.
However it may cause problems in the future, this change intends to address those in advance.

cf http://stackoverflow.com/questions/458550/standard-way-to-embed-version-into-python-package