Belatedly bump version number
This is in support of the practice of having the version number
in the repository be what it will be upon the next release.  In
particular, the first commit following a release should have the
version number bumped.  In a forthcoming merge, infrastructure for
appending commit hashes to version strings for non-release commits
will be implemented.