Version suffix of -dev-{%H,Unknown} for non-release installations
The primary mechanism is a file named commit_hash.txt under tulip/
If it is empty, then the installation is regarded as a release, so
should have a regular version string.  If it is exists and is nonempty,
then its contents are taken to be the commit hash from which the
installation was made, and "-dev-%H" is appended to the version string,
where %H is the hash.  Otherwise (i.e., commit_hash.txt is missing),
"-dev-Unknown" is appended to the version string.

The manner by which tulip/version.py is imported in setup.py is
improved in this changeset.  It now relies on the standard imp package.