bugfix: copy default values of labeled graph annotations
to avoid making labels of different nodes (or edges) point to the same object in memory (e.g., set()) when initializing them to the default values.
If that happened, then any direct change, for example adding a new proposition to some label: ts.states[state]['ap'].add('p'), would affect the labels of all other nodes (or edges) as well.