rename method LabeledDiGraph.trim_dead_states -> remove_deadends
remove is common for networkx methods

LabeledDiGraph should better be agnostic of "states" and use the term "nodes", because it is a networkx derivative. Although States is one of its attributes, it mostly wraps its functionality.
LabeledDiGraph may be split in the future into a pure-networkx subclass and a system-specific class (States and Transitions will reside in the latter only).