rm LabeledDiGraph.states.current
because it does not serve any good purpose.
If the user wants one, let them define such an attribute.

Moreover, it may be:
  - a set for generators
  - or a single state for transducers
because the semantics are different.

This indicates that it is a temporary value
that an algorithm (semantics) needs to keep track
of a run or a search (for an acceptor, or while
extracting a deterministic transducer from a
non-deterministic one), instead of part of the
definition of the object.

In any case it leads to implicit,
instead of explicit code, thus problematic.