extract synchronous product FTS * BA to algorithms module
because they just bloat the automata and transys modules.
They are non-trivial algorithms whose semantics are emphasized if the user has to call them as functions.
They operate on objects from classes defined in different modules, so wrapping them by methods leads to circular imports (if we want type checking).