rm unimplemented methods from FiniteStateMachine
because is_blocking and is_receptive are not very useful.
Receptiveness may not be required and depends on assumptions about the environment.

Products of machines are practically never taken explicitly.
They will be needed only for composing machines for simulation.
In that case, the semantics become more explicit by implementing them as functions.