shift matplotlib (and other graphics) imports from top to inline
in order to significantly accelerate loading of tulip.
The import time for long runs (e.g., discretization) may not be a significant portion of the total runtime.
However, for everyday development, typically characterized by many short runs, the time required to import matplotlib is painfully long.

Note the convention adopted from now on for inline imports, in order to reconcile them with pep8:
Inline imports are listed in comments below normal imports at the top of the file.
This offers the best of both worlds, and helps quickly inspect a file for imported dependencies (contrast with searching for "import" statements throughout the file, which is the alternative).

Related work:
  f11bf19765b707b2e3861d819ed887b9912d763a
  675216d05ffb733d2c6d1c1326b439414e0be000
  7e05aeac30d3d24db2cbc00154c5b80ad2bc2743