add new function: synth.synthesize_many (temp name)
that accepts a dict of transition systems, and so has a simpler signature.

This approach is explicit (better dict than list or set) and uses the "owner" attribute of each ts to decide whether the state is controlled by env or sys.
The explicit naming avoids conflicts of the integer variable names used to represent the current state in GR(1).

Instead of the arbitrary "loc", now the name is always ts_name + "_state", where ts_name is the key of that ts in the dict.
Therefore conflicts are avoided and the resulting var names are human readable (and not arbitrary, like "loc_1", "loc_2", … .

The option to model states with Booleans is not exposed any more, because it is inefficient.
In the future, it will be possible to use names of states directly in the GRSpec argument and those will be automatically replaced by the proper integer variable assignment, e.g., "mysys_state=2".

An additional improvement is that "ignore_init" and "bool_actions" are now just sets of ts names.