merge class DFA with NFA and name it FiniteWordAutomaton
to be explicit, and because a DFA is a subclass of an NFA, whose proper name should minimally include "finite word", to avoid confusion (there exists enough inconsistency of terminology in the literature).

An FWA data structure does not differ (at least for now) from a BA, but their semantics are different, so better separate them.
This becomes apparent if in the future a method "to_ltl" is added to infinite-word automata.
Of course whether this should be a method or a function is another debate and since it involves interfaces, e.g.,to ltl2ba, that will live elsewhere, such a method would cause cyclic imports, so it seems appropriate that it should be a function in another tulip sub package (transys doesn't know anything about formulas).

Also, rm unimplemented is_accepted method.