rename AST classes, removing the prefix AST
because it is redundant, reduces readability and increases clutter.
Instead, import the ast module and access its classes with the syntax: ast.Node
This is cleaner and also more explicit.
Note that "from . import ast" will shadow the builtin ast module. This is not a real problem because it can always be imported with a different name, if the python ast module must also be imported in the same namespace. In addition, their coexistence has not been necessary anywhere in tulip until now.