skip check in TypedDict if type is None
this allows relaxing somewhat the type checks. The state and edge attribute names are still checked, but the value not. It is a compromise to avoid the clumsy checks needed for symbolic attributes: too implicit.

It may be potentially better in the future to pull all the type checks into a method, to be run optionally if desired (and so avoid imposing constraints on the user, which is more pythonic, while still making the consistency checks available to be used).