mv conversion from a strategy to a transducer from interfaces.gr1c to synth
because state substitutions (arbitrary finite domains) rely on lists from within GRSpec. The elements of these lists are constrained to be such, that syntactic substitutions can be performed in GRSpec. By pulling this part back to synth, all information available in synth can be used, eventually allowing for integer states in transition systems, a priori substitutions (outside of GRSpec) and less fragility.