indent production rules to encapsulate them in parser class
because maintaining global state in the parser requires using either:

- global variables (less robust than creating a parser class)
- creating a parser class
- storing attributes in the parser's state which is made available to each production rule via its argument p, namely YaccProduction.parser. This approach is more implicit than using a parser class.

The parser class approach is also recommended by the PLY documentation, as "the easiest to manage" (with reference to lexer classes, but the same applies to parsers).

The state to be maintained is the networkx.DiGraph that will underly the AST in the future. A new graph need to be initialized for each formula parsed (so that the different trees returned will each have their own underlying nx graph instance).