add method: LTL_AST.sub_values
and allow skipping addition of node to graph in Node.__init__
Also, use id(self) at a single place for picking a unique integer for the node.
Store that in self.id for later reference. This allows for setting it to a value different than id(self).

Otherwise each AST node will be tied to a unique node in the underlying networkx.DiGraph, namely its id.
Replacing nodes will require removing that node and reconnecting a newly added node.
This is not necessary and by using self.id it can be avoided.