add method: LTL_AST.sub_constants
to substitute values of constants depending on the domain of the variable on the other side of the Binary operator above the Const. Note how substitution happens, which is interesting. From each constant, ascend until finding the first Binary operator, then descend on the other branch until finding the first variable (some more error handling for malformed formulas doesn't hurt to be added in the future). Then use the domain of that variable to decide with what integer to replace the string constant.