add env_vars, sys_vars attributes to strategy returned by load_aut_xml
because it is basic information deduced by the loader and fundamental for expanding the contracted strategy to a strategy on the game graph (that is a deterministic game graph, unless the strategy has some redundancy that allows the system more than one winning reactions at certain points in the game).

Motivation was unit testing.