add check for java as required dependency in setup.py
because for example on ubuntu systems java is not installed by default. gr1c is only checked as an optional dependency. jtlv is distributed together with tulip, but w/o a check for java either during installation or inside interfaces.jtlv, it is not so clear for the user what caused synthesis to fail.