rewrite ast
The main motivation is to use AST nodes as vertices of networkx.DiGraph.
Using integers as vertices and maintaining the map between the two is
cumbersome and fragile. It is also slower.

The benefit of using networkx comes from easy AST manipulation, like
attaching a subtree, or making in-place node substitutions.
So if more AST-related functionality is needed in the future,
manipulating the AST will be easier and lexical workarounds less
attractive.

By default, user-defined classes (as all the AST node types are) are
hashable, with __hash__ == __id__. So they can be used as nodes, w/o
modification. This allows removing the "id" attribute. However, the
"graph" attribute needs to remain, because otherwise the node classes
cannot access their neighbors. Using the __hash__ but leaving a "graph"
attribute that ties a node to some specific tree leads to a
contradiction.

So the "graph" attribute must be removed, if the AST nodes are to be
used as immutable. Deciding to remove "graph" necessitates rewriting the
classes to extract the tree flattening algorithm from them.