TRIV: Delete trailing whitespace
Commit fb4b59729b060a61e545b8f3cfb23ad524acc449 introduced whitespace
on blank lines into run_tests.py.  The original motivation was likely
to create uniform indentation, but the current preferred style is
to regard this as trailing whitespace.  As such, I am now deleting it.

Relevant discussion about coding style occurred on the tulip-control-discuss
mailing list around 6 Oct on a thread with subject "style".