add slugs flattener to AST
because slugs recognizes only capitalized literals for Boolean constants. In the future, the flatteners should be separated from the core AST classes, as extensions, in order to avoid spec.ast and spec.form to become god objects. Chaining the language transformations, instead of integrating them into single modules will be more robust and easier to debug.