rm method: GRSpec.sym_to_prop
because it is superseded by the syntactic manipulation utilities that are safer to use.
It replaces symbols using a lexical search with regular expressions, which does not distinguish between string literals and variable names. Working on the AST is safer.

Another motivation is to reduce confusion as to which method the user should pick (because the AST manipulation methods cover this functionality, for example GRSpec.sub_values).