mv language translation backends to the new module: `spec.translation`
to reduce the complexity of GRSpec and make it independent of specific tools. Tulip parsers LTL, it stores GR(1) specs in `GRSpec`, uses syntactic manipulation on a common AST (to be updated in one of the following commits), and when a translation is need, this AST is mapped to another ("isomorphic") AST that has different `flatten` methods, suited for the target language.

This also liberates from the Babylonian tendency of having differently named flatteners for each language (`to_gr1c`, `to_jtlv`, …). They make the data structures more difficult to deal with, violating the recommendation by Alan Perlis to have a large number of functions that operate on a few, fundamental data structures.

Each language's AST nodes have a different operator map attribute that they inherit from their parent abstract base class `Node`. So to produce subclasses that inherit from the updated parent `Node`, a node factory function will be defined in `spec.ast` and is used in `spec.translation` to create a node factory for each language.

The refactored `spec` package has a clear separation between the parser and AST manipulation frontend, and the flattening (translation) backends for the various languages. This makes it easier to add new formats w/o the danger to affect existing ones, as well as import from tulip in some other package and reuse the base AST node classes and factories to create new ones.

Also note that `spec.translation` is using some syntactic AST manipulation utilities from the `spec.transformation` module (to be created).