extract syntactic AST manipulation functions and Tree class to new module `spec.transformation`
because they are separate utilities that operate on AST graphs and need no be mixed with the modules that define AST node classes, nor GR(1) specifications. They are independent, general-purpose utilities.

The class `spec.ast.LTL_AST` has been renamed to `spec.transformation.Tree`, and significantly restructured (for example it does not need to worry about translations any more).

Note that the methods `LTL_AST.sub_values` and `LTL_AST.sub_constants` have been moved to functions of the same name in `spec.transformation`.

The `Tree` class has methods `from_recursive_ast` and `to_recursive_ast` for conversions from and to recursively defined AST (see below).

Rationale
=========
Note that as of 89920b64bb9ddc5d6946c7bd79f053e875cfabf3 the AST node classes formed a recursively defined tree, not a graph data structure as those in `networkx`, i.e., one that contains easily-accessible information for traversing it in either direction w/o too much extra code.

Those were converted to the current ones, that are only used as nodes in a `networkx` graph (user defined classes are hashable with their `id` as their `__hash__`). The reason was to make syntactic manipulations easier (since they occur several times throughout processing of formulae). Having AST nodes that recursively define a tree (with attributes like `operand`, `left`, and `right`) and use those as nodes of a `networkx` graph was deemed confusing and error-prone. However, `networkx` graphs are slightly less efficient when it comes to flattening, and some operations are still easier with the recursive definition, especially defining new transformations.

So for the benefit of simplicity, uniformity and to encourage resuse of the AST node factories, the AST node classes will be converted back to a recursive definition. The syntactic manipulations are significantly cleaner by using a `networkx` graph for traversing and modifying the tree.

For this reason, before syntactic manipulation, a `networkx` graph is constructed over a recursively defined AST, to be used as a "scaffold" for changes. During those changes, the underlying recursively defined AST is not touched (i.e., attributes like `operand`, `left`, `right` of the nodes belonging to the input tree are read-only). When the transformations are complete, the `networkx` scaffold is used to produce a new recursively defined AST, i.e., fresh node objects are created, and connected based on the transformed `networkx` graph (i.e., their attributes like `operand`, `left`, and `right` are set based on the edges of the final `networkx` graph).

This makes manipulation clean, and separate from everything else.
If one prefers to write a function that operates directly on the recursive AST, perhaps because that particular operation does not need to traverse the tree in both directions, then they can simply not use the module `spec.transformation`.